Create primeAddition
You are given an even number N and you need to represent the given number as the sum of primes. The prime numbers do not necessarily have to be distinct. It is guaranteed that at least one possible solution exists.
You need to determine the minimum number of prime numbers needed to represent the given number.